Add Spearman's correlation, and fix some assumptions for the student's t-test.
For the student's t-test, assume that the two samples of the dependent
variable are encompased by a single, continuous dependent variable off
of the combined test information. Only the continuous-continuous test
cases should expect the independent and dependent variable properties
to be set.

Additionally, utilize a property that specifies whether two samples are
paired or not in the t-test, u-test, and paired t-test.