Add constraints for paired t test.
Adapt constraints to match expectations for student's t test,
specifically that the independent variable is categorical with
two categories, and then the continuous results for each group
are both contained in the dependent variable. Additionally, for
bivariate tests the independent variable should have exactly two
groups.

Add a stub to CombinedData to represent whether the difference
between paired values is normally distributed.