Work towards printing assumptions and normality test results.
There is an issue where if 'hypothesize' is called multiple times in a
row, the property names printed in the assumptions get stale.

- Introduce a namedtuple for the results of the normality test so we can
nicely print the results.
- Move ResultData to a new file to resolve avoid a circular dependency
between evaulate_data_structures and solver.
- For each test in ResultData, get its assumptions from the test's
properties. If a property also has a "property test result" that means
we did some work to verify that property, so include that result in the
assumptions output.
- When printing the results for a test, print the assumptions before the result.
- In solver.py, keep track of the results of the normality test along with
whether we consider it normal based on the alpha value. If checking the
property returns a tuple, attach the second element to the property because
it holds the normality test results.
- If we don't check whether a property is true because it was a user
assumption, attach that assumption to the property.