Fix interpretation for Welch's t-test.
Feed the variables into the t-test in the correct order (left side of
the prediction, then right side of the prediction) so that the resulting
statistic has the sign that we are expecting when interpreting the result.
This ordering matches what we do for the independent t-test.

I moved setting the prediction before iterating through the categories,
and then I added comparing the categories to the prediction during the iteration.